Fixed warning

+++ game/src/player.rs
@@ -12,7 +12,7 @@ use n64_math::Vec2;
 const PLAYTER_START_POS: Vec2 = Vec2::new(0.5, 0.8);
 const SHIP_SPEED: f32 = 0.35;
 const SHIP_SHOOT_DELAY_MS: i32 = 150;
-pub const SHIP_SIZE: Vec2 = Vec2::new(32.0 / 320.0 as f32, 32.0 / 240.0 as f32);
+pub const SHIP_SIZE: Vec2 = Vec2::new(32.0 / 320.0, 32.0 / 240.0);
 
 pub struct Player {
     entity: OwnedEntity,
